Reorganized dll output linking to make the final build directory cleaner. Dlls will be split between a 3rdparty folder and a Libs folder.
Reorganized codebase to create two new Dlls to help with future updating so that end-user does not need to re-download the entire program for small changes.

Calculator.DLL : All files within the Calculator project are one large partial class to help prevent creating one large file.

DataClases.DLL : The following files have been moved to this DLL class library :
     Ability.cs,
     AbilityLibrary.cs,
     CharacterLibrary.cs,
     CharacterSheet.cs,
     CompanionSheet.cs,
     SheetExportSettings.cs,
     StatRules.cs,
     ZipExportSettings.cs,
     AudioPlayer.cs,
     ExcelManager.cs,
     ExceptionHandles.cs,
     Globals.cs,
     ProgramSettings.cs

Synovian Character Maker.exe : Updated all remaining files to link and reference new namespaces for Calculator.DLL and DataClases.DLL
          CharacterMaker.cs : Disabled functionality for import/export of txt and zips
          SheetExportSettingsForm.cs : Created new window to setup various settings for exporting character sheets.
          Program.cs : Main program class will now store object instances rather than static classes of the ExcelManager & Calculator. For DLL linking, added ResolveEventHandler() to handle all events in which the program fails to find a library.

+++ Synovian Character Maker/Forms/CharacterMaker/CharacterMaker.cs
@@ -4,8 +4,9 @@ using System.ComponentModel;
 using System.Linq;
 using System.IO;
 using System.Windows.Forms;
-using Synovian_Character_Maker.Data_Classes;
-using Synovian_Character_Maker.Static_Classes;
+using Synovian_Character_Maker.DataClasses;
+using Synovian_Character_Maker.DataClasses.Instanced;
+using Synovian_Character_Maker.DataClasses.Static;
 using Synovian_Character_Maker.Forms.CharacterMaker.CompanionMaker;
 
 namespace Synovian_Character_Maker.Forms.CharacterMaker
@@ -34,7 +35,7 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
         {
             if(saveCurrentOne)
             {
-                DataWriter.WriteCharacterToDiskTxt(current_characterSheet);
+                Program.excelManager.ExportSheet(current_characterSheet,"", new SheetExportSettings());
             }
 
             current_characterSheet = characterSheet;
@@ -201,7 +202,7 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
                 else if (extension == "xls" || extension == "xlsx")
                 {
                     //current_characterSheet = DataReader.LoadExelSheet(openCharacterDialog.FileName, ((extension == ".xls") ? IronXL.ExcelFileFormat.XLS : IronXL.ExcelFileFormat.XLSX));
-                    current_characterSheet = ExcelManager.ImportSheet(openCharacterDialog.FileName);
+                    current_characterSheet = Program.excelManager.ImportSheet(openCharacterDialog.FileName);
                 }
                 else
                 {
@@ -240,7 +241,7 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
             if (zipExportOptions.exitStatus == ZipExportOptions.ExitStatus.Failed)
                 return;
 
-            DataWriter.WriteCharacterToDiskZip(current_characterSheet, saveZip.FileName, zipExportOptions.exportSettings);
+            //DataWriter.WriteCharacterToDiskZip(current_characterSheet, saveZip.FileName, zipExportOptions.exportSettings);
             WriteLog("Saved character as zip to disk.");
         }
 
@@ -274,11 +275,12 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
 
         private void calculateButton_Click(object sender, EventArgs e)
         {
+            /*
             // Get the rules for the specific rank setup for this character.
-            int FeatPointMax = Program.statRules.FeatPoints[current_characterSheet.Rank];
-            int SkillPointsMax = Program.statRules.SkillPoints[current_characterSheet.Rank];
-            int SchoolsMax = Program.statRules.Schools[current_characterSheet.Rank];
-            int MasteriesMax = Program.statRules.Masteries[current_characterSheet.Rank];
+            int FeatPointMax = Program._statRules.FeatPoints[current_characterSheet.Rank];
+            int SkillPointsMax = Program._statRules.SkillPoints[current_characterSheet.Rank];
+            int SchoolsMax = Program._statRules.Schools[current_characterSheet.Rank];
+            int MasteriesMax = Program._statRules.Masteries[current_characterSheet.Rank];
 
             CalculatorLog calculatorLog = new CalculatorLog();
 
@@ -397,6 +399,24 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
                         else
                             continue;
                     }
+                }
+            }
+            // Calculate number of masterys
+            {
+                int numMaxMasteries = Program._statRules.Masteries[current_characterSheet.Rank];
+                int numMasteries = 0;
+
+                foreach(KeyValuePair<int,Ability_Mastery> pair in current_characterSheet.abilityMasteryDictionary)
+                {
+                    if (pair.Value >= Ability_Mastery.Mastery_Master) numMasteries++;
+                }
+
+
+                if(numMasteries > numMaxMasteries)
+                {
+                    valid = false;
+                    calculatorLog.AddToLog($"You {numMaxMasteries - numMasteries} more masteries than what you can currently have. You're max is {numMaxMasteries}");
+                    numErrors++;
                 }
             }
             // Calculate all prereq checks
@@ -603,7 +623,7 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
                                 numErrors++;
                             }
                         }
-                        else if (ability.Name == "Blaster" || ability.Name == "Slug Thrower" || ability.Name.Contains("Anti-") || ability.Name == "Covering Fire" || ability.Name == "Tactical Reload" || ability.Name == "Rail Shot" || ability.Name == "Laze Target" || ability.Name == "High Impact Bolt")
+                        else if (ability.Name == "Blaster" || ability.Name == "Slug Thrower" || ability.Name.Contains("Anti-Material") || ability.Name.Contains("Anti-Organic") || ability.Name.Contains("Anti-Armor") || ability.Name == "Covering Fire" || ability.Name == "Tactical Reload" || ability.Name == "Rail Shot" || ability.Name == "Laze Target" || ability.Name == "High Impact Bolt")
                         {
                             string[] rangeFeats = { "Blaster Training", "Marksman Training", "Pistol Training", "Heavy Weapons Training" };
                             bool hasFeat = false;
@@ -912,10 +932,15 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
             //    }
             //
             //    SkipCompaionChecks:;                
-            //}
+            //}*/
 
-            calculatorLog.AddToLog($"Results are in: your sheet is {((valid == true) ? "valid" : "invalid")}.");
-            calculatorLog.AddToLog($"Number of errors: {numErrors}");
+            CalculatorLog calculatorLog = new CalculatorLog();
+            string log = "";
+            int errors = 0;
+            bool success = Program.calculator.Calculate(ref current_characterSheet, ref log, ref errors);
+            calculatorLog.AddToLog(log);
+            calculatorLog.AddToLog($"Results are in: your sheet is {((success == true) ? "valid" : "invalid")}.");
+            calculatorLog.AddToLog($"Number of errors: {errors}");
             calculatorLog.ShowDialog();
             WriteLog("Did big brain math.");
         }
@@ -931,7 +956,14 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
 
         private void saveExcel_FileOk(object sender, CancelEventArgs e)
         {
-            DataWriter.ExportCharacterSheetExcel(current_characterSheet, saveExcel.FileName, (saveExcel.FileName.Split('.')[1] == "xlsx") ? DataWriter.ExcelFormats.XLSX : DataWriter.ExcelFormats.XLS);
+            return;
+
+            SheetExportSettingsForm sheetExportSettingsForm = new SheetExportSettingsForm();
+            sheetExportSettingsForm.ShowDialog();
+            SheetExportSettings sheetExportSettings = sheetExportSettingsForm.sheetExportSettings;
+
+            //DataWriter.ExportCharacterSheetExcel(current_characterSheet, saveExcel.FileName, sheetExportSettings, (saveExcel.FileName.Split('.')[1] == "xlsx") ? DataWriter.ExcelFormats.XLSX : DataWriter.ExcelFormats.XLS);
+            Program.excelManager.ExportSheet(current_characterSheet, saveExcel.FileName, sheetExportSettings);
             WriteLog("Saved character to disk as excel");
         }
 
@@ -983,10 +1015,25 @@ namespace Synovian_Character_Maker.Forms.CharacterMaker
 
         private void googleDriveButton_Click(object sender, EventArgs e)
         {
+            SheetExportSettingsForm sheetExportSettings = new SheetExportSettingsForm();
+            sheetExportSettings.ShowDialog();
+            if (!sheetExportSettings.finishedSetup) return;
+            SheetExportSettings sheetExport = sheetExportSettings.sheetExportSettings;
+
             if (!Directory.Exists(Globals.TempFolder)) Directory.CreateDirectory(Globals.TempFolder);
-            DataWriter.ExportCharacterSheetExcel(current_characterSheet, $"{Globals.TempFolder}\\{current_characterSheet.Name}.xlsx", DataWriter.ExcelFormats.XLSX);
-            Static_Classes.Networking.GoogleDrive.GoogleDriveManager.SubmitSheetToDrive($"{Globals.TempFolder}\\{current_characterSheet.Name}.xlsx");
+
+
+            // Create the file
+            //DataWriter.ExportCharacterSheetExcel(current_characterSheet, $"{Globals.TempFolder}\\{current_characterSheet.Name}.xlsx", sheetExport, DataWriter.ExcelFormats.XLSX);
+            Program.excelManager.ExportSheet(current_characterSheet, $"{Globals.TempFolder}\\{current_characterSheet.Name}.xlsx", sheetExport);
+
+            // Send file to google drive
+            Networking.GoogleDrive.GoogleDriveManager.SubmitSheetToDrive($"{Globals.TempFolder}\\{current_characterSheet.Name}.xlsx");
+
+            // Delete the file, we don't need it on disk
             File.Delete($"{Globals.TempFolder}\\{current_characterSheet.Name}.xlsx");
+
+
             WriteLog("Sent sheet to google drive.");
         }
